insight_api/requirements.txt
fastapi==0.104.1
python-dotenv==1.0.0
uvicorn==0.24.0.post1
